requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
Click==7.0
durable-rules==2.0.14
experta==1.9.4
frozendict==1.2
isort==4.3.21
lazy-object-proxy==0.0.0
mccabe==0.6.1
pathspec==0.6.0
pylint==2.4.4
regex==2019.11.1
schema==0.6.7
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2
